# openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "API MiniMarket Plus",
        "description": "Microservicio backend de MiniMarket Plus. Expone la gestion de\nproductos, categorias, inventario, carrito de compras, ventas,\ndetalle de ventas y usuarios del sistema.\n\nEsta documentacion se genera con OpenAPI Specification (OAS 3)\nmediante springdoc-openapi, asegurando la consistencia entre el\ncodigo y la documentacion de la API.",
        "contact": {
            "name": "Equipo Backend - MiniMarket Plus",
            "email": "soporte@minimarketplus.cl"
        },
        "license": {
            "name": "Uso academico Duoc UC",
            "url": "https://www.duoc.cl"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Entorno de desarrollo local"
        }
    ],
    "security": [
        {
            "autenticacionBasica": []
        }
    ],
...
    "paths": {
...
        "/api/productos/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener producto por ID",
                "description": "Devuelve la informacion detallada de un producto segun su identificador.",
                "operationId": "obtenerProductoPorId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador del producto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Producto"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar producto",
                "description": "Actualiza los datos de un producto existente identificado por su ID.",
                "operationId": "actualizarProducto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador del producto a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Producto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "401": {
                        "description": "No autenticado"
                    },
                    "200": {
                        "description": "Producto actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Producto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar producto",
                "description": "Elimina un producto del catalogo por su identificador.",
                "operationId": "eliminarProducto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador del producto a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Producto eliminado (sin contenido)"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            }
        },
...
        "/api/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Listar productos",
                "description": "Obtiene la lista completa de productos del catalogo.",
                "operationId": "listarProductos",
                "responses": {
                    "200": {
                        "description": "Lista de productos obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Producto"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear producto",
                "description": "Registra un nuevo producto en el catalogo. Requiere una categoria existente.",
                "operationId": "guardarProducto",
                "requestBody": {
                    "description": "Datos del producto a crear",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Producto"
                            },
                            "examples": {
                                "Producto nuevo": {
                                    "description": "Producto nuevo",
                                    "value": {
                                        "nombre": "Leche 1L",
                                        "precio": 1200.0,
                                        "stock": 30,
                                        "categoria": {
                                            "id": 1
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "401": {
                        "description": "No autenticado"
                    },
                    "200": {
                        "description": "Producto creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Producto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Categoria": {
                "required": [
                    "nombre"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Identificador unico de la categoria.",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "description": "Nombre unico de la categoria.",
                        "example": "Bebidas"
                    },
                    "productos": {
                        "type": "array",
                        "description": "Productos asociados a esta categoria.",
                        "readOnly": true,
                        "items": {
                            "$ref": "#/components/schemas/Producto"
                        }
                    }
                },
                "description": "Categoria que agrupa productos del catalogo (por ejemplo Bebidas o Abarrotes)."
            },
            "Producto": {
                "required": [
                    "categoria",
                    "nombre",
                    "precio",
                    "stock"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Identificador unico del producto.",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "description": "Nombre comercial del producto.",
                        "example": "Agua Mineral 1L"
                    },
                    "precio": {
                        "type": "number",
                        "description": "Precio de venta unitario en pesos chilenos.",
                        "format": "double",
                        "example": 890.0
                    },
                    "stock": {
                        "type": "integer",
                        "description": "Cantidad disponible en stock.",
                        "format": "int32",
                        "example": 50
                    },
                    "categoria": {
                        "$ref": "#/components/schemas/Categoria"
                    }
                },
                "description": "Producto disponible en el catalogo de MiniMarket Plus."
            },
...
        },
        "securitySchemes": {
            "autenticacionBasica": {
                "type": "http",
                "description": "Autenticacion HTTP Basic. Use el usuario y la contrasena semilla (admin / admin123) en el boton Authorize para consumir los endpoints protegidos.",
                "name": "autenticacionBasica",
                "scheme": "basic"
            }
        }
    }
}